Default behaviour when no reset type available
Supermicro Redfish API implementation does not return any available
values in the reset type field. However, it supports `ForceRestart`
but it does not support `PushPowerButton` that is the default
behaviour in the `fillForOn` and `fillForOff` functions.

To make the implementation work for the Supermicro implementation,
the default case has been changed to set the action to
`ForceRestart`. A new case for `PushPowerButton` has been added
so that systems stating that they support that functionality will
not see any change of behaviour.

Supermicro Redfish Guide p.21-22:
https://www.supermicro.com/manuals/other/RedfishRefGuide.pdf